refactor(compiler): Move out unused deps lint
