+++ src/cargo/core/compiler/mod.rs
@@ -51,7 +51,7 @@ pub mod timings;
 mod unit;
 pub mod unit_dependencies;
 pub mod unit_graph;
-mod unused_deps;
+pub mod unused_deps;
 
 use std::borrow::Cow;
 use std::cell::OnceCell;
