+++ src/cargo/diagnostics/rules/unused_dependencies.rs
@@ -1,5 +1,6 @@
 use std::path::Path;
 
+use cargo_util_schemas::manifest;
 use cargo_util_schemas::manifest::TomlPackageBuild;
 use cargo_util_terminal::report::AnnotationKind;
 use cargo_util_terminal::report::Group;
@@ -7,15 +8,22 @@ use cargo_util_terminal::report::Level;
 use cargo_util_terminal::report::Origin;
 use cargo_util_terminal::report::Patch;
 use cargo_util_terminal::report::Snippet;
-use tracing::instrument;
+use tracing::{debug, instrument, trace};
 
 use super::STYLE;
 use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::Package;
+use crate::core::PackageId;
 use crate::core::Workspace;
+use crate::core::compiler::BuildContext;
+use crate::core::compiler::BuildRunner;
+use crate::core::compiler::Unit;
+use crate::core::compiler::unused_deps::UnusedDepState;
+use crate::core::dependency::DepKind;
 use crate::diagnostics::DiagnosticStats;
 use crate::diagnostics::Lint;
+use crate::diagnostics::LintLevel;
 use crate::diagnostics::LintLevelProduct;
 use crate::diagnostics::get_key_value_span;
 use crate::diagnostics::rel_cwd_manifest_path;
@@ -157,4 +165,223 @@ pub(crate) fn lint_package(
     }
 
     Ok(())
+}
+
+#[instrument(skip_all)]
+pub fn lint_build_results(
+    build_runner: &BuildRunner<'_, '_>,
+    warn_count: &mut usize,
+    error_count: &mut usize,
+) -> CargoResult<()> {
+    for (pkg_id, states) in &build_runner.unused_dep_state.states {
+        let Some(pkg) = get_package(&build_runner.unused_dep_state, pkg_id) else {
+            continue;
+        };
+        let toml_lints = pkg
+            .manifest()
+            .normalized_toml()
+            .lints
+            .clone()
+            .map(|lints| lints.lints)
+            .unwrap_or(manifest::TomlLints::default());
+        let cargo_lints = toml_lints
+            .get("cargo")
+            .cloned()
+            .unwrap_or(manifest::TomlToolLints::default());
+        let LintLevelProduct {
+            level: lint_level,
+            source,
+        } = LINT.level(
+            &cargo_lints,
+            pkg.rust_version(),
+            pkg.manifest().unstable_features(),
+        );
+
+        if lint_level == LintLevel::Allow {
+            for (dep_kind, state) in states.iter() {
+                for ext in state.unused_externs.iter().flatten() {
+                    debug!(
+                        "pkg {} v{} ({dep_kind:?}): ignoring unused extern `{ext}`, lint is allowed",
+                        pkg_id.name(),
+                        pkg_id.version(),
+                    );
+                }
+            }
+            continue;
+        }
+
+        let manifest_path = rel_cwd_manifest_path(pkg.manifest_path(), build_runner.bcx.gctx);
+        let mut lint_count = 0;
+        for (dep_kind, state) in states.iter() {
+            for ext in state.unused_externs.iter().flatten() {
+                let mut used_in_dev = false;
+                match dep_kind {
+                    DepKind::Normal => {
+                        if let Some(state) = states.get(&DepKind::Development)
+                            && state
+                                .unused_externs
+                                .as_ref()
+                                .is_some_and(|ue| !ue.contains(ext))
+                        {
+                            used_in_dev = true;
+                        }
+                    }
+                    DepKind::Development => {
+                        if let Some(state) = states.get(&DepKind::Normal)
+                            && state.externs.contains_key(ext)
+                        {
+                            trace!(
+                                "pkg {} v{} ({dep_kind:?}): ignoring unused extern `{ext}`, inherited from normal dependency",
+                                pkg_id.name(),
+                                pkg_id.version(),
+                            );
+                            continue;
+                        }
+                    }
+                    DepKind::Build => {}
+                }
+                let Some(extern_state) = state.externs.get(ext) else {
+                    // not one we care to report
+                    debug!(
+                        "pkg {} v{} ({dep_kind:?}): ignoring unused extern `{ext}`, untracked dependent",
+                        pkg_id.name(),
+                        pkg_id.version(),
+                    );
+                    continue;
+                };
+                if state.seen_units.len() != state.needed_units {
+                    debug_assert_ne!(state.externs.len(), 0, "assumes tracked is checked first");
+                    // Some compilations errored without printing the unused externs.
+                    // Don't print the warning in order to reduce false positive
+                    // spam during errors.
+                    debug!(
+                        "pkg {} v{} ({dep_kind:?}): ignoring unused extern `{ext}`, {} outstanding units",
+                        pkg_id.name(),
+                        pkg_id.version(),
+                        state.needed_units - state.seen_units.len()
+                    );
+                    continue;
+                }
+                if is_transitive_dep(&extern_state.unit, &state.seen_units, build_runner.bcx) {
+                    debug!(
+                        "pkg {} v{} ({dep_kind:?}): ignoring unused extern `{ext}`, may be activating features",
+                        pkg_id.name(),
+                        pkg_id.version(),
+                    );
+                    continue;
+                }
+
+                // Implicitly added dependencies (in the same crate) aren't interesting
+                let dependency = if let Some(dependency) = &extern_state.manifest_deps {
+                    dependency
+                } else {
+                    continue;
+                };
+                for dependency in dependency {
+                    let manifest = pkg.manifest();
+                    let document = manifest.document();
+                    let contents = manifest.contents();
+                    let level = lint_level.to_diagnostic_level();
+                    let emitted_source = LINT.emitted_source(lint_level, source);
+                    let toml_path = dependency.toml_path();
+
+                    let mut primary = Group::with_title(level.primary_title(LINT.desc));
+                    if let Some(document) = document
+                        && let Some(contents) = contents
+                        && let Some(span) = get_key_value_span(document, &toml_path)
+                    {
+                        let span = span.key.start..span.value.end;
+                        primary = primary.element(
+                            Snippet::source(contents)
+                                .path(&manifest_path)
+                                .annotation(AnnotationKind::Primary.span(span)),
+                        );
+                    } else {
+                        primary = primary.element(Origin::path(&manifest_path));
+                    }
+                    if lint_count == 0 {
+                        primary = primary.element(Level::NOTE.message(emitted_source));
+                    }
+                    lint_count += 1;
+                    let mut report = vec![primary];
+                    if let Some(document) = document
+                        && let Some(contents) = contents
+                        && let Some(span) = get_key_value_span(document, &toml_path)
+                    {
+                        let span = span.key.start..span.value.end;
+                        let mut help =
+                            Group::with_title(Level::HELP.secondary_title("remove the dependency"));
+                        help = help.element(
+                            Snippet::source(contents)
+                                .path(&manifest_path)
+                                .patch(Patch::new(span, "")),
+                        );
+                        report.push(help);
+                    }
+                    if used_in_dev {
+                        let help = Group::with_title(Level::HELP.secondary_title(
+                            "to still use for development builds, move to `dev-dependencies`",
+                        ));
+                        report.push(help);
+                    }
+
+                    if lint_level.is_warn() {
+                        *warn_count += 1;
+                    }
+                    if lint_level.is_error() {
+                        *error_count += 1;
+                    }
+                    build_runner
+                        .bcx
+                        .gctx
+                        .shell()
+                        .print_report(&report, lint_level.force())?;
+                }
+            }
+        }
+    }
+    Ok(())
+}
+
+fn get_package<'s>(
+    unused_dep_state: &'s UnusedDepState,
+    pkg_id: &PackageId,
+) -> Option<&'s Package> {
+    let state = unused_dep_state.states.get(pkg_id)?;
+    let mut iter = state.values();
+    let state = iter.next()?;
+    let mut iter = state.seen_units.iter();
+    let unit = iter.next()?;
+    Some(&unit.pkg)
+}
+
+#[instrument(skip_all)]
+fn is_transitive_dep(
+    direct_dep_unit: &Unit,
+    seen_units: &Vec<Unit>,
+    bcx: &BuildContext<'_, '_>,
+) -> bool {
+    let mut queue = std::collections::VecDeque::new();
+    for root_unit in seen_units {
+        for unit_dep in &bcx.unit_graph[root_unit] {
+            if root_unit.pkg.package_id() == unit_dep.unit.pkg.package_id() {
+                continue;
+            }
+            if unit_dep.unit == *direct_dep_unit {
+                continue;
+            }
+            queue.push_back(&unit_dep.unit);
+        }
+    }
+
+    while let Some(dep_unit) = queue.pop_front() {
+        for unit_dep in &bcx.unit_graph[dep_unit] {
+            if unit_dep.unit == *direct_dep_unit {
+                return true;
+            }
+            queue.push_back(&unit_dep.unit);
+        }
+    }
+
+    false
 }
